Готово

+++ AdmissionCommittee.Manager/StudentsManager.cs
@@ -39,7 +39,7 @@ namespace AdmissionCommittee.Manager
             }
         }
 
-        async Task IStudentsManager.AddStudentAsync(Student student)
+        async Task IStudentsManager.AddStudentAsync(Student student, CancellationToken cancellationToken)
         {
             var sw = Stopwatch.StartNew();
             try
@@ -54,7 +54,7 @@ namespace AdmissionCommittee.Manager
             }
         }
 
-        async Task IStudentsManager.UpdateStudentAsync(Student student)
+        async Task IStudentsManager.UpdateStudentAsync(Student student, CancellationToken cancellationToken)
         {
             var sw = Stopwatch.StartNew();
 
@@ -70,13 +70,20 @@ namespace AdmissionCommittee.Manager
             }
         }
 
-        async Task IStudentsManager.DeleteStudentAsync(Guid id)
+        async Task<Student?> IStudentsManager.DeleteStudentAsync(Guid id, CancellationToken cancellationToken)
         {
             var sw = Stopwatch.StartNew();
 
             try
             {
-                await studentService.DeleteStudentAsync(id);
+                var student = await studentService.GetStudentByIdAsync(id);
+
+                if (student != null)
+                {
+                    await studentService.DeleteStudentAsync(id);
+                }
+
+                return student;
             }
             finally
             {
@@ -86,7 +93,7 @@ namespace AdmissionCommittee.Manager
             }
         }
 
-        async Task<Student?> IStudentsManager.GetStudentByIdAsync(Guid id)
+        async Task<Student?> IStudentsManager.GetStudentByIdAsync(Guid id, CancellationToken cancellationToken)
         {
             var sw = Stopwatch.StartNew();
